delete 6 exercises
