+++ filter/ex2.js
@@ -1,22 +1,30 @@
-/* Array.prototype.filter - Exercice 2
+/* Array.prototype.filter - Exercice 3
 
-Ecrire une fonction getMultiplesOf qui reçoit deux arguments:
-1. un tableau de nombres.
-2. un nombre.
-En utilisant filter, elle doit renvoyer un tableau ne contenant que les
-nombres du 1er argument qui sont multiples du nombre passé en 2ème.
+Ecrire une fonction keepStarks, qui prend en arguments un tableau de chaînes
+(plus spécifiquement, des noms de personnages de Game of Thrones).
+
+La fonction doit renvoyer un tableau ne contenant que les membres de la famille Stark.
+
+Il y a plusieurs façons d'écrire cette fonction, et outre filter, tu auras probablement
+besoin d'une de ces méthodes de String: endsWith ou split.
 
 Exemple d'entrée:
-  1er argument:  [-12, -10, -7, -1, 5, 8, 11, 15, 20, 27]
-  2ème argument: 5
+  [
+    'Bran Stark',
+    'Cersei Lannister',
+    'Sandor Clegane',
+    'Arya Stark',
+    'Yara Greyjoy',
+    'Sansa Stark'
+  ]
 
-  Sortie attendue:
-  [-10, 5, 15, 20]
+Sortie attendue:
+  ['Bran Stark', 'Arya Stark', 'Sansa Stark']
 
  */
 
-function getMultiplesOf(numbers, n) {
+function keepStarks(names) {
 }
 
 // Ne pas modifier l'export
-module.exports = getMultiplesOf;
+module.exports = keepStarks;
