+++ tests/filter2.test.js
@@ -1,10 +1,17 @@
-const getMultiplesOf = require('../filter/ex2.js');
+const keepStarks = require('../filter/ex2.js');
 
 xdescribe('filter - exercice 2', () => {
 
-  test('getMultiplesOf', () => {
-    expect(getMultiplesOf([-12, -10, -7, -1, 5, 8, 11, 15, 20, 27], 5))
-    .toEqual([-10, 5, 15, 20]);
+  test('keepStarks', () => {
+    expect(keepStarks([
+      'Bran Stark',
+      'Cersei Lannister',
+      'Sandor Clegane',
+      'Arya Stark',
+      'Yara Greyjoy',
+      'Sansa Stark'
+    ]))
+    .toEqual(['Bran Stark', 'Arya Stark', 'Sansa Stark']);
   });
 
 });
