+++ tests/filter3.test.js
@@ -1,17 +1,59 @@
-const keepStarks = require('../filter/ex3.js');
+const filterOnPrice = require('../filter/ex3.js');
+
+const products = [
+  {
+    name: 'iPhone X',
+    price: 1159
+  },
+  {
+    name: 'Xiaomi Mi A2',
+    price: 209
+  },
+  {
+    name: 'Samsung Galaxy Note 9',
+    price: 992
+  },
+  {
+    name: 'Huawei P20',
+    price: 480
+  },
+  {
+    name: 'Huawei P20 Pro',
+    price: 649
+  }
+];
 
 xdescribe('filter - exercice 3', () => {
 
-  test('keepStarks', () => {
-    expect(keepStarks([
-      'Bran Stark',
-      'Cersei Lannister',
-      'Sandor Clegane',
-      'Arya Stark',
-      'Yara Greyjoy',
-      'Sansa Stark'
-    ]))
-    .toEqual(['Bran Stark', 'Arya Stark', 'Sansa Stark']);
+  test('filterOnPrice (price <= 500)', () => {
+    expect(filterOnPrice(products, 500))
+    .toEqual([
+      {
+        name: 'Xiaomi Mi A2',
+        price: 209
+      },
+      {
+        name: 'Huawei P20',
+        price: 480
+      }
+    ]);
+  });
+  test('filterOnPrice (price <= 649)', () => {
+    expect(filterOnPrice(products, 649))
+    .toEqual([
+      {
+        name: 'Xiaomi Mi A2',
+        price: 209
+      },
+      {
+        name: 'Huawei P20',
+        price: 480
+      },
+      {
+        name: 'Huawei P20 Pro',
+        price: 649
+      }
+    ]);
   });
 
 });
